+++ src/main/java/wooteco/chess/SparkChessApplication.java
@@ -3,12 +3,12 @@ package wooteco.chess;
 import static spark.Spark.*;
 
 import wooteco.chess.controller.SparkChessController;
+import wooteco.chess.repository.DataSource;
 import wooteco.chess.repository.GameDAO;
-import wooteco.chess.repository.JDBCGameDAO;
+import wooteco.chess.repository.MySQLDataSource;
+import wooteco.chess.repository.jdbc.JDBCGameDAO;
+import wooteco.chess.repository.jdbc.JDBCTemplate;
 import wooteco.chess.service.GameService;
-import wooteco.chess.utils.jdbc.DataSource;
-import wooteco.chess.utils.jdbc.JDBCTemplate;
-import wooteco.chess.utils.jdbc.MySQLDataSource;
 
 public class SparkChessApplication {
 	public static void main(String[] args) {
