+++ src/main/java/wooteco/chess/domain/game/KingCatchFinished.java
@@ -15,7 +15,6 @@ public class KingCatchFinished extends Finished {
 		}
 	}
 
-	//움직이고 왕이 잡혔는지 확인 한 뒤, 잡혔으면 턴 교체 없이 바로 끝 !
 	@Override
 	public Team getWinner() {
 		return turn;
